Fix problems uncovered by compiling on alpha:  * define int2p() and p2int(), and use them rather than casts  * define p2str(), and use it rather than "%x" to print pointers  * ensure some printf arguments match their formats, by adding casts
bk: 3f7cba2842dWod65Avt_WnPSRAf9Aw

+++ src/bkd_http.c
@@ -266,7 +266,7 @@ navbutton(int active, int tag, char *start, char *end)
 		if (arguments[ct-1] == '|') --ct;
 
 		sprintf(buf, "href=\"%.*s?%.*s\">",
-		    sep-start, start, ct, arguments);
+		    (int)(sep-start), start, ct, arguments);
 	} else {
 		sprintf(buf, "href=\"%.*s\">", sep-start,start);
 	}
@@ -275,7 +275,7 @@ navbutton(int active, int tag, char *start, char *end)
 		   : "<font size=2 color=yellow>");
 
 	if (sep < end) {
-		sprintf(buf, "%.*s", end-sep, sep);
+		sprintf(buf, "%.*s", (int)(end-sep), sep);
 		out(buf);
 		start = 0;
 	} else if (strneq(start,"ChangeSet", 9)) {
@@ -347,7 +347,7 @@ navbutton(int active, int tag, char *start, char *end)
 		}
 	}
 	if (start) {
-		sprintf(buf, "%.*s", sep-start, start);
+		sprintf(buf, "%.*s", (int)(sep-start), start);
 		out(buf);
 	}
 	out("</font></a>\n");
